actualización

+++ src/main/java/com/jecsamtech/kapebackend/dto/ActualizarCantidadDTO.java
@@ -11,5 +11,5 @@ public class ActualizarCantidadDTO {
 
     @NotNull(message = "La cantidad es obligatoria")
     @Min(value = 1, message = "La cantidad debe ser al menos 1")
-    private Integer cantidad;
+    private Long cantidad;
 }
