+++ src/main/java/com/jecsamtech/kapebackend/dto/CarritoHasCafeDTO.java
@@ -17,5 +17,5 @@ public class CarritoHasCafeDTO {
 
     @NotNull(message = "La cantidad es obligatoria")
     @Min(value = 1, message = "La cantidad debe ser al menos 1")
-    private Integer cantidad;
+    private Long cantidad;
 }
